About to start audio call implementation

+++ models/notification.js
@@ -1,6 +1,7 @@
 const mongoose = require('mongoose');
 const helper = require('../helpers/encrypt');
-
+const { newError } = require('../helpers/error');
+const status = require('http-status');
 const notificationSchema = new mongoose.Schema(
   {
     sender: {
@@ -34,16 +35,63 @@ notificationSchema.statics.findFriendRequests = async function (req) {
   }).populate('sender', '-password');
 };
 
+notificationSchema.statics.getSentFriendRequests = async function (req) {
+  const requests = await this.find({
+    sender: req.user.id,
+    notification_type: 'friend_request',
+  })
+    .select('receiver ')
+    .populate('receiver', '-password -socketId -updated_at');
+
+  return requests;
+};
+
+notificationSchema.statics.receivedFriendRequests = async function (req) {
+  const requests = await this.find({
+    receiver: req.user.id,
+    notification_type: 'friend_request',
+  })
+    .select('sender')
+    .populate('sender', '-password -socketId -updated_at');
+
+  return requests;
+};
+
+notificationSchema.statics.deleteSentFriendRequest = async function (req) {
+  const deleted = await this.findOneAndRemove({
+    _id: req.params.id,
+    sender: req.user.id,
+  });
+  return deleted._id;
+};
+
 notificationSchema.statics.create = async function (req) {
-  return await Notification({
+  const existingNotification = await this.findOne({
     sender: req.user.id,
     receiver: req.body.receiver,
-    notification_type: req.body.notification_type,
-  }).save();
+  });
+
+  if (!existingNotification) {
+    return await Notification({
+      sender: req.user.id,
+      receiver: req.body.receiver,
+      notification_type: req.body.notification_type,
+    }).save();
+  } else {
+    throw newError('Notification already sent!', status.BAD_REQUEST);
+  }
 };
 
 notificationSchema.statics.delete = async function (id) {
   const deleted = await this.findByIdAndDelete(id);
+  return deleted;
+};
+
+notificationSchema.statics.deleteFriendRequest = async function (id) {
+  const deleted = await this.findOneAndRemove({
+    _id: id,
+    receiver: req.user.id,
+  });
   return deleted;
 };
 
@@ -56,8 +104,10 @@ notificationSchema.statics.deleteAfterAcceptingFriendRequest = async function (
     receiver: receiver,
   });
 
-  const deleted = await this.findByIdAndDelete(notification._id);
-  return deleted;
+  if (notification) {
+    const deleted = await this.findByIdAndDelete(notification._id);
+    return deleted;
+  }
 };
 
 const Notification = mongoose.model('Notification', notificationSchema);
